Split registry into subpackage
